pyproject.toml
[project]
name = "forecastcam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flet>=0.82.2",
    "google-generativeai>=0.8.6",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "statsmodels>=0.14.6",
]
